requirements.txt
streamlit>=1.30.0
yt-dlp>=2024.1.0
openai>=1.0.0
python-dotenv>=1.0.0
